Reorganize domaing
