+++ src/components/Account/AccountList.jsx
@@ -2,7 +2,7 @@ import React, { Component } from "react";
 import { connect } from "react-redux";
 import LoadingComponent from "../App/LoadingComponent";
 import AccountSummary from "./AccountSummary";
-import { fetchAccounts } from "./accountActions";
+import { fetchAccounts } from "../../domain/account/actions";
 import "./AccountList.css";
 
 class AccountList extends Component {
