+++ src/domain/async/actions.js
@@ -1,8 +1,4 @@
-import {
-  ASYNC_ACTION_START,
-  ASYNC_ACTION_FINISH,
-  ASYNC_ACTION_ERROR
-} from "./asyncConstants";
+import { ASYNC_ACTION_START, ASYNC_ACTION_FINISH, ASYNC_ACTION_ERROR } from "./constants";
 
 export const asyncActionStart = () => {
   return {
