+++ src/store/index.js
@@ -1,7 +1,7 @@
 import { combineReducers } from "redux";
-import accountReducers from "../components/Account/accountReducers";
-import authReducers from "../components/Auth/authReducers";
-import asyncReducer from "../components/Async/asyncReducer";
+import accountReducers from "../domain/account/reducers";
+import authReducers from "../domain/auth/reducers";
+import asyncReducer from "../domain/async/reducers";
 
 const rootReducer = combineReducers({ async: asyncReducer, auth: authReducers, accounts: accountReducers });
